+++ Changes.md
@@ -12,6 +12,7 @@ API
 
 - Widget : Added `currentButtons()` static method. This returns the state of the mouse buttons during the last UI event to be processed.
 - LazyMethod : Added `deferUntilButtonRelease` option.
+- MetadataAlgo : Added `firstViewableAncestor()` function.
 
 1.6.12.0 (relative to 1.6.11.1)
 ========